requirements.txt
beautifulsoup4==4.13.5
blinker==1.9.0
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
Flask==3.1.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
python-dotenv==1.1.1
reportlab==4.4.4
requests==2.32.5
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
